chore(cosmic-theme): move Roundness to libcosmic
